requirements.txt
python-telegram-bot==20.7
sqlite3
python-dotenv==1.0.0
flask==2.0.1
flask-cors==3.0.10
pyTelegramBotAPI==4.12.0
